research_assistant/requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# RAG & Vector DB
chromadb>=0.5.0
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
langchain-text-splitters>=0.2.0
pypdf>=4.2.0
numpy<2.0.0

# For OpenAI (PAID)
openai>=1.0.0
langchain-openai>=0.1.0

transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0